==> OrganizeData/ERGuarantee/step02_sort_er_guarantee_data_KX.py <==
#!/usr/bin/env python

# -*- coding: utf-8 -*-
# @Filename: step02_sort_er_guarantee_data_KX
# @Date: 2024/8/17
# @Author: WONG Kar Xiong
# @Email: wangyouan@gamil.com

# -*- coding: utf-8 -*-
"""
Created on Mon Aug 12 02:02:23 2024

@author: User
"""
import pandas as pd
import numpy as np

# =============================================================================

df1 = pd.read_csv("C:/Users/User/Desktop/ER/2005_2023_ER_guarantee/ER_Guarantee.csv", error_bad_lines=False)

df1['SignDate'] = pd.to_datetime(df1['SignDate'], errors='coerce')
df1['StartDate'] = pd.to_datetime(df1['StartDate'], errors='coerce')
df1['DeclareDate'] = pd.to_datetime(df1['DeclareDate'], errors='coerce')

# STEP 1
# 行政区划
Internal = """
北京
东城 西城 海淀 朝阳 丰台 门头沟 石景山 房山 通州 顺义 昌平 大兴 怀柔 平谷 延庆 密云区
上海
黄浦  徐汇  长宁  静安  普陀  虹口  杨浦  闵行  宝山  嘉定  浦东新  金山  松江  青浦  奉贤  崇明  
天津
河东  河西  南开  河北  红桥  东丽  西青  津南  北辰  武清  宝坻  滨海新  宁河  静海  蓟州  
重庆 
万州  黔江  涪陵  渝中  大渡口  江北  沙坪坝  九龙坡  南岸  北碚  渝北  巴南  长寿  江津  合川  永川  南川  綦江  大足  璧山  铜梁  潼南  荣昌  开州  梁平  武隆  城口  丰都  垫江  忠  云阳  奉节  巫山  巫溪  石柱 自治  秀山  自治  酉阳  自治  彭水  自治  
深圳  
福田  罗湖  盐田  南山  宝安  龙岗  龙华  坪山  光明  大鹏新  
山东 
济南    中  历下   天桥   槐荫   历城   长清   章丘   平阴   济阳   商河 
青岛    南    北   城阳   四方   李沧   黄岛   崂山   胶南   胶州   平度   莱西   即墨 
淄博   张店   临淄   淄川   博山   周村   桓台   高青   沂源 
枣庄    中   山亭   峄城   台儿庄   薛城   滕州 
东营   东营   河口   垦利   广饶   利津 
烟台   芝罘   福山   牟平   莱山   龙口   莱阳   莱州   招远   蓬莱   栖霞   海阳   长岛 
潍坊   潍城   寒亭   坊子   奎文   青州   诸城   寿光   安丘   高密   昌邑   昌乐   临朐 
济宁    市中   任城   曲阜   兖州   邹城   鱼台   金乡    嘉祥   微山   汶上   泗水   梁山 
泰安   泰山   岱岳   新泰   肥城   宁阳   东平 
威海   环翠   乳山   文登   荣成 
日照   东港   岚山   五莲   莒 
莱芜   莱城   钢城 
临沂   兰山   罗庄   河东   沂南   郯城   沂水   苍山   费   平邑   莒南   蒙阴   临沭 
德州   德城   乐陵   禹城   陵   宁津   齐河   武城   庆云   平原   夏津   临邑 
聊城   东昌府   临清   高唐   阳谷   茌平   莘   东阿   冠 
滨州   滨城   邹平   沾化   惠民   博兴   阳信   无棣 
菏泽   牡丹   鄄城   单   郓城   曹   定陶   巨野   东明   成武  
河北   
石家庄  井陉  正定  行唐  灵寿  高邑  深泽  赞皇  无极  平山  元氏  赵  晋州  新乐  辛集 。
保定  涞水  阜平  定兴  唐  高阳  容城  涞源  望都  安新  易  曲阳  蠡  顺平  博野  雄  安国  高碑店  涿州  定州 。
唐山  滦南  乐亭  迁西  玉田  遵化  迁安  滦州 
秦皇岛  青龙   昌黎  卢龙 
邯郸  临漳  成安  大名  涉  磁  邱  鸡泽  广平  馆陶  曲周  武安  魏 
邢台  邢台  临城  内丘  柏   隆尧  任  南和  巨鹿  新河  广宗  平   威  清河  临西  南宫  沙河  宁晋 
张家口  张北  康保  沽源  尚义  蔚  阳原  怀安  涿鹿  赤城  怀来 
承德  承德  兴隆  滦平  隆化  丰宁   宽城   围场    平泉 
沧州  沧  青  东光  海兴  盐山  肃宁  南皮  吴桥  献  孟村   泊头  黄骅  河间  任丘 
廊坊  固安  永清  香河  大城  文安  大厂   霸州  三河 
衡水  枣强  武邑  武强  饶阳  安平  故城  阜城  深州  景  
黑龙江   
哈尔滨  道里  南岗  道外  平房  松北  香坊  呼兰  阿城  双城  依兰  方正  宾  巴彦  木兰  通河  延寿  尚志  五常 
齐齐哈尔  龙沙  建华  铁锋  富拉尔基  昂昂溪  碾子山  梅里斯   讷河  甘南  龙江  克山  克东  依安  拜泉  泰来  富裕 
牡丹江  东安  西安  爱民  阳明  绥芬河  海林  宁安  穆棱  东宁  林口 
佳木斯  富锦  同江  桦南  桦川  汤原  抚远  前进  向阳  东风  郊 
大庆  萨尔图  龙凤  让胡路  红岗  大同  肇州  肇源  林甸  杜尔伯特  
鸡西  密山  虎林  鸡东  鸡冠  恒山  滴道  城子河  梨树  麻山 
双鸭山  尖山  岭东  四方台  宝山  集贤  宝清  友谊  饶河 
伊春  伊美  乌翠  友好  金林  铁力  嘉荫  汤旺  丰林  南岔  大箐山 
七台河  桃山  新兴  茄子河  勃利 
鹤岗  绥滨  萝北  东山  兴安  向阳  工农  南山  兴山 
黑河  爱辉  北安  五大连池  嫩江  逊克  孙吴 
绥化  北林  肇东  安达  海伦  兰西  庆安  绥棱  望奎  明水  青冈 
大兴安岭  加格达奇  松岭  呼中  新林  漠河  呼玛  塔河  
辽宁 
长春  南关  宽城  朝阳  二道  绿园  双阳  九台  农安  榆树  德惠  公主岭 
吉林  昌邑  龙潭  船营  丰满  永吉  蛟河  桦甸  舒兰  磐石 
四平  铁西  铁东  梨树  伊通   双辽 
辽源  龙山  西安  东丰  东辽 
通化  东昌  二道江  通化  辉南  柳河  集安 
白山  浑江  江源  抚松  靖宇  长白   临江 
松原  宁江  前郭尔罗斯   长岭  乾安  扶余 
白城  洮北   赉  通榆  洮南  大安 
延边   延吉  图们  敦化  珲春  龙井  和龙  汪清  安图 
长白山管委会 长白山国家级自然保护 管理局 安图 安图长白山旅游经济开发  长白山和平旅游度假  抚松 抚松长白山旅游经济开发  长白 长白山南坡旅游经济园 
梅河口 吉林 直管 级  
辽宁  
沈阳  和平  沈河  大东  皇姑  铁西  苏家屯  浑南  沈北新  于洪  辽中  新民  康平  法库 
大连  中山  西岗  沙河口  甘井子  旅顺口  金州  普兰店  瓦房店  庄河  长海 
鞍山  铁东  铁西  立山  千山  海城  台安  岫岩  
抚顺  新抚  顺城  望花  东洲  抚顺  新宾   清原  
本溪  平山  溪湖  明山  南芬  本溪   桓仁  
丹东  元宝  振兴  振安  东港  凤城  宽甸  
锦州  古塔  凌河  太和  凌海  北   黑山  义 
营口  站前  西   鲅鱼圈  老边  盖州  大石桥 
阜新  海州  新邱  太平  清河门  细河  阜新   彰武 
辽阳  白塔  文圣  宏伟  弓长岭  太子河  灯塔  辽阳 
盘锦  双台子  兴隆台  大洼  盘山 
铁岭  银州  清河  调兵山  开原  铁岭  昌图  西丰 朝阳  双塔  龙城  北票  凌源  朝阳  建平  喀喇沁左翼  
葫芦岛  连山  龙港  南票  兴城  绥中  建昌  
内蒙古 
呼和浩特  回民  新城  玉泉  赛罕  土默特左旗 托克托  和林格尔  武川  清水河 	
包头  昆都仑  东河  青山  石拐  九原  白云鄂博矿  土默特右旗 固阳  达尔罕茂明安联合旗	
乌海  海勃湾  海南  乌达 	
赤峰  红山  元宝山  松山  阿鲁科尔沁旗 巴林左旗 巴林右旗 林西  克什克腾旗 翁牛特旗 喀喇沁旗 宁城  敖汉旗	
通辽  科尔沁  霍林郭勒  科尔沁左翼中旗 科尔沁左翼后旗 开鲁  库伦旗 奈曼旗 扎鲁特旗	
鄂尔多斯  康巴什  东胜  准格尔旗 达拉特旗 鄂托克前旗 鄂托克旗 杭锦旗 乌审旗 伊金霍洛旗 
呼伦贝尔  海拉尔  满洲里 （代管扎赉诺尔 ） 牙克石  扎兰屯  额尔古纳  根河  阿荣旗 鄂伦春  莫力达瓦      陈巴尔虎旗 新巴尔虎左旗 新巴尔虎右旗
巴彦淖尔  临河  五原  磴口  乌拉特前旗 乌拉特中旗 乌拉特后旗 杭锦后旗	
乌兰察布  集宁  丰   卓资  化德  商都  兴和  凉城  察哈尔右翼前旗 察哈尔右翼中旗 察哈尔右翼后旗 四子王旗	
兴安盟 乌兰浩特  阿尔山  科尔沁右翼前旗 科尔沁右翼中旗 扎赉特旗 突泉 	
锡林郭勒盟 锡林浩特  二连浩特  阿巴嘎旗 苏尼特左旗 苏尼特右旗 东乌珠穆沁旗 西乌珠穆沁旗 太仆寺旗 镶黄旗 正镶白旗 正蓝旗 多伦 	
阿拉善盟	 阿拉善左旗 阿拉善右旗 额济纳旗 
山西   
太原  小店  迎泽  杏花岭  尖草坪  万柏林  晋源  清徐  阳曲  娄烦  古交 
大同  新荣  平城  云冈  云州  阳高  天   广灵  灵丘  浑源  左云 
朔州  朔城  平鲁  山阴  应  右玉  怀仁 
忻州  忻府  定襄  五台  代  繁峙  宁武  静乐  神池  五寨  岢岚  河曲  保德  偏关  原平 
阳泉  城  矿  郊  平定  盂 
吕梁  离石  文水  交城  兴  临  柳林  石楼  岚  方山  中阳  交口  孝义  汾阳 
晋中  榆次  太谷  榆社  左权  和顺  昔阳  寿阳  祁  平遥  灵石  介休 
长治  潞州  上党  屯留  潞城  襄垣  平顺  黎城  壶关  长子  武   沁  沁源 
晋城  城  沁水  阳城  陵川  泽州  高平 
临汾  尧都  曲沃  翼城  襄汾  洪洞  古  安泽  浮山  吉   宁  大宁  隰  永和  蒲  汾西  侯马  霍州 
运城  盐湖  临猗  万荣  闻喜  稷山  新绛  绛  垣曲  夏  平陆  芮城  永济  河津  
陕西 
西安  蓝田   周至  户  高陵 
宝鸡   凤翔  岐山  扶风  眉  陇  千阳  麟游  凤  太白 
咸阳  泾阳  乾  礼泉  永寿  彬  长武  旬邑  淳化  武功 
渭南  华  潼关  大荔  蒲城  澄城  白水  合阳  富平 
铜川  宜君 
延安  延长  延川  子长  安塞  志丹  吴起  甘泉  富  洛川  宜川  黄龙  黄陵 
榆林  神木  府谷  横山  靖边  定边  绥德  米脂  佳  吴堡  清涧  子洲 
汉中  南郑  城固  洋   西   勉   宁强  略阳   巴  留坝  佛坪 
安康  石泉  汉阴  宁陕  紫阳  岚皋  平利   坪  白河 
商洛  洛南  丹凤  商南  山阳   安  柞水  
宁夏  
银川  石嘴山  吴忠  固原  中卫  永宁  贺兰  平罗  同心  盐池  西吉  隆德  泾源  彭阳  中宁  海原  
甘肃 
兰州  城关  七里河  西固  安宁  红古  永登  榆中  皋兰 
嘉峪关 
金昌  金川  永昌 
白银  白银  平川  会宁  靖远  景泰 
天水  秦州  麦积  甘谷  武山  秦安  清水  张家川  
武威  凉州  民勤  古浪 和天祝  
张掖  甘州  民乐  临泽  高台  山丹  肃南  
平凉  崆峒  泾川  灵台  崇信  华亭  庄浪  静宁 
酒泉  肃州  玉门  敦煌  金塔  瓜州  肃北  阿克塞  
庆阳  西峰  庆城 华池 宁   原 合水 正宁 环 
定西  安定  通渭  陇西  渭源  临洮  漳  岷 
陇南  武都  成  文  宕昌  康  西和  礼  徽  两当 
临夏   临夏  临夏  永靖  广河  和政  康乐     积石山    
甘南   合作  夏河 碌曲 玛曲 迭部 舟曲 临潭 卓尼  
新疆维吾尔 
乌鲁木齐  水磨沟  天山  沙依巴克  新   头屯河  达坂城  米东  乌鲁木齐 	
克拉玛依  克拉玛依  独山子  白碱滩  乌尔禾 	
吐鲁番 	 高昌  鄯善  托克逊 	
哈密  伊州  伊吾  巴里坤哈萨克 	
阿克苏  阿克苏  温宿  库车  沙雅  新和  拜城  乌什  阿瓦提  柯坪 	
喀什  喀什  疏附  疏勒  英吉沙  泽普  莎车  叶城  麦盖提  岳普湖  伽师  巴楚  塔什库尔干塔吉克 	
和田  和田  和田  墨玉  皮山  洛浦  策勒  于田  民丰 	
昌吉   昌吉  阜康  呼图壁  玛纳斯  奇台  吉木萨尔  木垒哈萨克 	
博尔塔拉蒙古  博乐  阿拉山口  精河  温泉 	
巴音郭楞蒙古  库尔勒  焉耆   轮台  尉犁  若羌  且末  和静  和硕  博湖 	
克孜勒苏柯尔克孜  阿图什  阿克陶  阿合奇  乌恰 	
伊犁哈萨克  伊宁  奎屯  霍尔果斯  尼勒克  伊宁  霍城  巩留  新源  昭苏  特克斯  察布查尔锡伯 	
塔城 	 塔城  乌苏  沙湾  额敏  托里  裕民  和布克赛尔蒙古 	
阿勒泰  阿勒泰  布尔津  富蕴  福海  哈巴河  青河  吉木乃 	
新疆生产建设兵团 石河子  阿拉尔  图木舒克  五家渠  北屯  铁门关  双河  可克达拉  昆玉  胡杨河  新星  
青海  
西宁  城中  城东  城西  城北  湟中  大通    湟源 
海东  乐都  平安  民和    互助   化隆   循化  
海北   海晏  祁连  刚察  门源  	
黄南   同仁  尖扎  泽库  河南  
海南   共和  同德  贵德  兴海  贵南 	
果洛   玛沁  班玛  甘德  达日  久治  玛多 	
玉树   玉树  杂多  称多  治多  囊谦  曲麻莱 
海西    德令哈  格尔木  茫崖  天峻  都兰  乌兰  大柴旦  
西藏  
拉萨  林周 当雄 达孜 曲水 墨竹工卡 尼木 堆龙德庆 旁多
那曲	 黑河 巴青 班戈 比如 安多 聂荣 索  达木萨迦 申扎
昌都	 昌都 江达 贡觉 宁静 洛隆 丁青 边坝 察雅 桑昂曲 八宿 类乌齐 左贡
山南	 乃东 桑日 加查 朗  拉加里 隆子 错那 哲古 穷结 洛扎 贡嘎 扎囊
江孜	 江孜 仁布 浪卡子 打隆 白朗 亚东
日喀则 日喀则 南木林 定结 萨迦 拉孜 定日 聂拉木 吉隆 萨嘎 昂仁 谢通门
阿里	 噶尔 日土 革吉 改则 仲巴 普兰 札达
林芝	 林芝 墨脱 工布江达 雪巴 波密 米林 嘉黎 
河南 
郑州  中原  二七  管城  金水  上街  惠济  巩义  荥阳  新密  新郑  登封  中牟 
开封  鼓楼  龙亭  顺河  禹王台 金明  杞  通许  尉氏  开封  兰考 
洛阳  西工  老城  瀍河  涧西  吉利  洛龙  偃师  孟津  新安  栾川  嵩  汝阳  宜阳  洛宁  伊川 
平顶山  新华  卫东  湛河  石龙  舞钢  汝州  宝丰  叶  鲁山  郏 
焦作  山阳  解放  中站  马村  沁阳  孟州  修武  博爱  武陟  温 
鹤壁  淇滨  山城  鹤山  浚  淇 
新   卫滨  红旗  凤泉  牧野  卫辉  辉   新   获嘉  原阳  延津  封丘  长垣 
安阳  北关  文峰  殷都  龙安  林州  安阳  汤阴  滑  内黄 
濮阳  华龙  清丰  南乐  范  台前  濮阳 
许昌  魏都  禹州  长葛  许昌  鄢陵  襄城 
漯河  源汇  郾城  召陵  舞阳  临颍 
三门峡  湖滨  义马  灵宝  渑池  陕  卢氏 
南阳  卧龙  宛城  邓州  南召  方城  西峡   平  内   淅川  社旗  唐河  新野  桐柏 
商丘  梁园  睢阳  永城  民权  睢  宁陵  柘城  虞城  夏邑 
信阳  浉河  平桥  罗山  光山  新  商城  固始  潢川  淮滨  息 
周口  川汇  项城  扶沟  西华  商水  沈丘  郸城  淮阳  太康  鹿邑 
驻马店  驿城  西平  上蔡  平舆  正阳  确山  泌阳  汝南  遂平  新蔡 
济源  
安徽  
合肥	 瑶海  庐阳  蜀山  包河  巢湖  肥东  肥西  长丰  庐江 
芜湖	 镜湖  弋江  鸠江  湾沚  繁昌   无为  南陵 
蚌埠 龙子湖  蚌山  禹会  淮上  五河  固   怀远 
淮南	 大通  田家庵  谢家集  八公山  潘集  凤台  寿 
马鞍山 	花山  雨山  博望  含山  和  当涂 
淮北	 相山  杜集  烈山  濉溪 
铜陵 铜官  郊  义安  枞阳 
安庆	 迎江  大观  宜秀  桐城  潜山  怀宁  太湖  宿松  望江  岳西 
黄山	 屯溪  黄山  徽州  歙  休宁  黟  祁门 
阜阳 颍州  颍泉  颍东  界首  颍上  临泉  阜南  太和 
宿州 埇桥  萧  砀山  灵璧  泗 
滁州	 琅琊  南谯  天长  明光  全椒  来安  凤阳  定远 
六安	 金安  裕安  叶集  霍邱  霍山  金寨  舒城 
宣城	 宣州  广德  宁国  郎溪  泾  绩溪  旌德 
池州	 贵池  青阳  石台  东至 
亳州	 谯城  蒙城  涡阳  利辛   江苏 下属  
南京  玄武  秦淮  建邺  鼓楼  浦口  栖霞  雨花台  江宁  六合  溧水  高淳  
无锡  滨湖  梁溪  新吴  锡山  惠山  江阴  宜兴 
徐州  云龙  鼓楼  贾汪  泉山  铜山  邳州  新沂  丰  沛  睢宁 
常州  天宁  钟楼  新北  武进  金坛  溧阳 
苏州  姑苏  虎丘  吴中  相城  吴江  昆山  常熟  张家港  太仓  
南通  崇川  通州  海门    启东  海安  如皋  如东 
连云港 	 海州  连云  赣榆  东海  灌云  灌南 
淮安  清江浦  淮安  淮阴  洪泽  涟水  盱眙  金湖 
盐城  亭湖  盐都  大丰  东台  响水  滨海  阜宁  射阳  建湖 
扬州  广陵  邗江  江都  仪征  高邮  宝应 
镇江  	京口  润州  丹徒  丹阳  扬中  句容 
泰州  海陵  高港  姜堰  兴化  靖江  泰兴 
宿迁  宿城  宿豫  沭阳  泗阳  泗洪  
湖北  
武汉  江岸  江汉  硚口  汉阳  武昌  青山  洪山  蔡甸  江夏  黄陂  新洲  东西湖  汉南 
黄石  黄石港  西塞山  下陆  铁山  大冶  阳新 
十堰 	 张湾  茅箭  郧阳  丹江口  郧西  竹山  竹溪  房 
宜昌  当阳  夷陵  西陵  伍家岗  点军  猇亭  宜都  枝江  远安  兴山  秭归  长阳   五峰  
襄阳  襄城  樊城  襄州  枣阳  宜城  老河口  南漳  保康  谷城 
鄂州  鄂城  华容  梁子湖 
荆门  沙洋  钟祥  京山  东宝  掇刀 
孝感  孝南  应城  安陆  汉川  云梦  孝昌  大悟 
荆州  荆州  沙   松滋  石首  洪湖  监利    江陵  公安 
黄冈  黄州  麻城  武穴  团风  浠水  罗田  英山  蕲春  黄梅  红安 
咸宁  咸安  赤壁  嘉鱼  通城  崇阳  通山 
随州  曾都  广水  随 
恩施    恩施  利川  建始  巴东  宣恩  咸丰  来凤  鹤峰 
仙桃  潜江  天门  神农架林  
湖南 
长沙  芙蓉  天心  岳麓  开福  雨花  望城  宁   浏阳  长沙 
株洲  天元  芦淞  荷塘  石峰  渌口  醴陵  攸  茶陵  炎陵 
湘潭  	雨湖  岳塘  湘   韶山  湘潭 
衡阳  雁峰  石鼓  珠晖  蒸湘  南岳  耒阳  常宁  衡阳  衡南  衡山  衡东  祁东 
邵阳  双清  大祥  北塔  武冈  邵东  新邵  邵阳  隆回  洞口  绥宁  新宁  城步  
岳阳 	 岳阳楼  云溪  君山  汨罗  临湘  湘阴  岳阳  华容  平江 
常德  武陵  鼎城  津   安   汉寿  桃源  临澧  石门  澧 
张家界 	 永定  武陵源  慈利  桑植 
益阳  资阳  赫山  沅江  安化  桃江  南  大通湖  
娄底  娄星  冷水江  涟源  双峰  新化 
郴州  北湖  苏仙  资兴  桂阳  宜章  永兴  嘉禾  临武  汝城  桂东  安仁 
永州  零陵  冷水滩  祁阳     双牌  东安  道  宁远  新田  蓝山  江永  江华  
怀化  	鹤城  洪江  洪江  中方  沅陵  辰溪  溆浦  会同  麻阳   新晃   芷江   靖州    通道  
湘西   	 吉首  泸溪  凤凰  花垣  保靖  古丈  永顺  龙山  
浙江
杭州	 上城  拱墅  西湖  滨江  余杭  临平  钱塘  萧山  富阳  临安  建德  桐庐  淳安 		
宁波	 海曙  江北  北仑   海  鄞州  奉化  余姚  慈溪  象山  宁海 		
温州	 鹿城  龙湾  瓯海  洞头  瑞安  乐清  龙港    永嘉  平阳  苍南  文成  泰顺 		
绍兴	 越城  柯桥  上虞  诸暨  嵊州  新昌 		
湖州	 吴兴  南浔  德清  长兴  安吉 		
嘉兴	 南湖  秀洲  海宁  平湖  桐   嘉善  海盐 		
金华	 婺城  金东  兰溪  东阳  永康  义乌  武义  浦江  磐安 		
衢州	 柯城  衢江  江山  常山  开化  龙游 		
台州	 椒江  黄岩  路桥  临海  温岭  玉环  三门  天台  仙居 		
丽水	 莲都  龙泉  青田  缙云  遂昌  松阳  云和  庆元  景宁  		
舟山	 定海  普陀  岱山  嵊泗  
江西 
南昌	 东湖  西湖  青云谱  青山湖  新建  红谷滩  南昌  安义  进贤 
九江	 浔阳  濂溪  柴桑  瑞昌  共青城  庐山  永修  德安  都昌  湖口  彭泽  武宁  修水 
上饶	 信州  广丰  广信   德兴  玉山  铅山  横峰  弋阳  余干  鄱阳  万年  婺源 
抚州	 临川  东   南城  黎川  南丰  崇仁  乐安  宜黄  金溪  资溪  广昌 
宜春	 袁州  高安  丰城  樟树  奉新  万载  上高  宜丰  靖安  铜鼓 
吉安	 吉州  青原  井冈山  吉安  吉水  新干  永丰  泰和  遂川  万安  安福  永新  峡江 
赣州	 章贡  南康  赣   龙南   瑞金  信丰  大余  定南  全南  寻乌  安远  宁都  于都  会昌  石城  上犹  兴国  崇义 
景德  珠山  昌江  乐平  浮梁 
萍  安源  湘东  莲花  上栗  芦溪 
新余	 渝水  分宜 
鹰潭	 月湖  余江  贵溪  
福建  
福州  闽侯  连江  罗源  闽清  永泰  平潭  福清 
漳州  云霄  漳浦  诏安  长泰  东山  南靖  平和  华安  龙海 
泉州  石狮  晋江  南安  惠安  安溪  永春  德化  金门 
三明  永安  明溪  清流  宁化  大田  尤溪  沙  将乐  泰宁  建宁 
莆田  仙游 
南平  邵武  武夷山  建瓯  顺昌  浦城  光泽  松溪  政和 
龙岩  漳平  长汀  上杭  武平  连城 
宁德  福安  福鼎  霞浦  古田  屏南  寿宁  周宁 
厦门  
广东  
广州  越秀  海珠  荔湾  天河  白云  黄埔  花都  番禺  南沙  从化  增城 
深圳  福田  罗湖  盐田  南山  宝安  龙岗  龙华  坪山  光明 
珠海  香洲  金湾  斗门 
汕头  	金平  龙湖  澄海  濠江  潮阳  潮南  南澳 
佛山  禅城  南海  顺德   高明  三水 
韶关  浈江  武江  曲江  乐昌  南雄  仁化  始兴  翁源  新丰  乳源  
河源  源城  东源  和平  龙川  紫金  连平 
梅州  	梅江  梅   兴宁  平远  蕉岭  大埔  丰顺  五华  
惠州  惠城  惠阳  惠东  博罗  龙门 
汕尾  城  陆丰  海丰  陆河 
东莞 
中山 
江门  蓬江  江海  新会  台山  开平  鹤山  恩平 
阳江  江城  阳东  阳春  阳西 
湛江  赤坎  霞山  麻章  坡头  雷州  廉江  吴川  遂溪  徐闻 
茂名  茂南  电白  信宜  高州  化州 
肇庆  	端州  鼎湖  高要    四会  广宁  德庆  封开  怀集 
清远  清城  清新  英德  连州  佛冈  阳山  连山    连南  
潮州  	湘桥  潮安  饶平 
揭阳  	榕城  揭东  普宁  揭西  惠来 
云浮  云城  云安  罗定  新兴  郁南  
广西  
南宁  兴宁  青秀  江南  西 塘  良庆  邕宁  武鸣  横州  隆安  马山  上林  宾阳 
柳州  	城中  鱼峰  柳南  柳北  柳江  柳城  鹿寨  融安  融水   三江  
桂林  秀峰  叠彩  象山  七星  雁山  临桂  荔浦  阳朔  灵川  全州  兴安  永福  灌阳  资源  平乐  龙胜各族  恭城  
梧州  万秀  长洲  龙圩  岑溪  苍梧  藤  蒙山 
北海  海城  银海  铁山港  合浦 
崇左  江州  凭祥  扶绥  宁明  龙州  大新  天等 
来宾  	兴宾  合山  忻城  象州  武宣  金秀  
贺州  八步  平桂  昭平  钟山  富川  
玉林  玉州  福绵  北流  容  陆川  博白  兴业 
百色  右江  田阳  靖西  平果  田东  德保  那坡  凌云  乐业  田林  西林  隆林各族 
河池  金城江  宜州  南丹  天峨  凤山  东兰  罗城   环江   巴马   都安   大化  
钦州  钦南  钦北  灵山  浦北 
防城港 	 港口  防城  东兴  上思 
贵港  港北  港南  覃塘  桂平  平南  
海南
海口  秀英 	秀英街道 海秀街道 长流  西秀  海秀  石山  永兴  东山 	
龙华  中山街道 滨海街道 大同街道 金贸街道 金宇街道 海垦街道 城西  龙桥  龙泉  新坡  遵谭 	
琼山  国兴街道 府城街道 凤翔街道 滨江街道 龙塘  云龙  红旗  旧州  三门坡  甲子  大坡 	
美兰  白龙街道 白沙街道 博爱街道 海甸街道 蓝天街道 海府路街道 人民路街道 和平南街道 新埠街道 灵山  演丰  三江  大致坡 和海口 三江农场	
三亚  海棠  吉阳  天涯  崖州 		
三沙  西沙  永兴（ ）  七连屿  永乐群岛  中沙群岛 其中黄岩岛由中国中央政府直接控制
南沙  南沙 	
儋州  三都街道 那大  南丰  雅星  和庆  大成  新洲  光村  东成  中和  峨蔓  兰洋  王五  排浦  海头  木棠  白马井 
五指山  通什  毛阳  番阳  南圣  毛道  水满  畅好 
文昌  文城  铺前  锦山  翁田  冯坡  抱罗  公坡  昌洒  龙楼  东郊  文教  东阁  潭牛  东路  蓬莱  重兴  会文 	
琼海  嘉积  长坡  万泉  潭门  阳江  中原  博鳌  塔洋  龙江  石壁  会山  大路 	
万宁  万城  北大  和乐  后安  长丰  大茂  礼纪  山根  三更罗  南桥  龙滚  东澳 	
东方  八所  感城  三家  板桥  四更  新龙  大田  东河  天安  江边 	
定安  定城  新竹  龙湖  雷鸣  龙门  岭口  翰林  龙河  黄竹  富文 	
屯昌  屯城  南坤  新兴  西昌  坡心  南吕  枫木  乌坡 	
澄迈  瑞溪  加乐  福山  桥头  金江  老城  永发  文儒  中兴  仁兴  大丰 	
临高  临城  东英  波莲  调楼  新盈  南宝  加来  和舍  多文  博厚  皇桐 	
白沙   牙叉  七坊  打安  邦溪  南开  元门  阜龙  细水  青松  金波  荣邦 	
昌江   石碌  昌化  海尾  七叉  十月田  乌烈  叉河  王下 	
乐东   抱由  千家  万冲  九所  利国  大安  志仲  尖峰  黄流  佛罗  莺歌海 	
陵水   椰林  新村  英州  本号  隆广  三才  光坡  文罗  黎安  提蒙  群英 	
保亭    新政  响水  什玲  保城  加茂  三道  南林  六弓  毛感 	
琼中    营根  黎母山  湾岭  和平  长征  中平  红毛  上安  什运  吊罗山  
贵州 
贵阳  清   遵义  赤水  仁怀  安顺  毕节  兴义  都匀  福泉  六盘水  关岭   铜仁  凯里 
开阳  修文  息烽  水城  盘  遵义  绥阳  桐梓  习水  凤冈  正安  余庆  湄潭  普定  德江  江口  思南  石阡  黔西  大方  织金    赫章  纳雍  望谟  兴仁  普安  册亨  晴隆  贞丰  安龙  施秉  从江  锦屏   远  麻江  台江  天柱  黄平  榕江  剑河  三穗  雷山  黎平  岑巩  丹寨  贵定  惠水  罗甸  瓮安  荔波  龙里  平塘  长顺  独山 
三都   松桃   玉屏   沿河   道真    务川   平坝   宁    紫云    印江    威宁     
四川  
成都  锦江  青羊  金牛  武侯  成华  青白江  龙泉驿  新都  温江  双流  郫都  金堂  大邑  蒲江  新津  都江堰  彭州  崇州  邛崃  简阳 
绵阳  涪城  游仙  安州  梓潼  三台  盐亭  平武  江油  北川  
自贡  自流井  贡井  大安  沿滩  荣  富顺 
攀枝花  东  西  仁和  米易  盐边 
泸州  江阳  龙马潭  纳溪  泸  合江  叙永  古蔺 
德阳  旌阳  罗江  广汉  什邡  绵竹  中江 
广元  利州  昭化  朝天  旺苍  青川  剑阁  苍溪 
遂宁  船山  安居  射洪  蓬溪  大英 
内江   中  东兴  资中  威远  隆昌 
乐山   中  沙湾  五通桥  金口河  犍为  井研  夹江  沐川  峨眉山  峨边   马边  
资阳  雁江  安岳  乐至 
宜宾  叙州  翠屏  南溪  江安  长宁  高  筠连  珙  兴文  屏山 
南充  顺庆  高坪  嘉陵  西充  南部  蓬安  营山  仪陇  阆中 
达州  通川  达川  宣汉  开江  大竹  渠  万源 
雅安  雨城  名山  荥经  汉源  石棉  天全  芦山  宝兴 
阿坝    马尔康  金川  小金  阿坝  若尔盖  红原  壤塘  汶川  理  茂  松潘  九寨沟  黑水 
甘孜   康定  泸定  丹巴  九龙  雅江  道孚  炉霍  甘孜  新龙  德格  白玉  石渠  色达  理塘  巴塘   城  稻城  得荣 
凉山   西昌  德昌  会理  会东  宁南  普格  布拖  昭觉  金阳  雷波  美姑  甘洛  越西  喜德  冕宁  盐源  木里  
广安  广安  前锋  邻水  武胜  岳池  华蓥 
巴中  巴州  恩阳  平昌  通江  南江 
眉山  东坡  彭山  仁寿  丹棱  青神  洪雅  
云南 
昆明  呈贡  盘龙  五华  官渡  西山  晋宁  东川  安宁  富民  宜良  嵩明  石林   禄劝    寻甸   
曲靖  	麒麟  沾益  马龙  宣威  富源  罗平  师宗  陆良  会泽 
玉溪  红塔  江川  澄江  通海  华宁  易门  峨山   新平    元江    
昭通  昭阳  水富  鲁甸  巧家  盐津  大关  永善  绥江   雄  彝良  威信 
保山  	隆阳  腾冲  施甸  龙陵  昌宁 
丽江  古城  永胜  华坪  玉龙   宁蒗  
普洱  思茅  宁洱    墨江   景东   景谷     沅     江城    孟连     澜沧   西盟  
临沧  临翔  凤庆  云  永德   康  双江      耿马    沧源  
德宏    	芒  瑞丽  梁河  盈江  陇川 
怒江  	 泸水  福贡  贡山    兰坪   
迪庆   香格里拉  德钦  维西  
大理   大理  祥云  宾川  弥渡  永平  云龙  洱源  剑川  鹤庆  漾濞   南涧   巍山   
楚雄   楚雄  双柏  牟定  南华  姚安  大姚  永仁  元谋  武定  禄丰 
红河    蒙自  个旧  开远  弥勒  建水  石屏  泸西  绿春  元阳  红河  金平     河口   屏边  
文山    文山  砚山  西畴  麻栗坡  马关  丘北  广南  富宁 
西双版纳   景洪  勐海  勐腊 
""".split()

# Create a boolean mask for rows that contain any of the keywords, check below
mask = df1['GuaranteeName'].str.contains('|'.join(Internal))

data = df1[~mask]     # Data without the Internal, most likely foreign, next step


# Step 2
# Define regex patterns
english_pattern = r'[A-Za-z]'
chinese_pattern = r'[\u4e00-\u9fff]'

# Identify rows with English characters
data_en = data[data['GuaranteeName'].str.contains(english_pattern, na=False)] #intl

# Identify rows with only Chinese characters
data_cn = data[~data['GuaranteeName'].str.contains(english_pattern, na=False) &
                data['GuaranteeName'].str.contains(chinese_pattern, na=False)]

# Identify global+cities+offshores
WW1 = """
南美
哥伦比亚	厄瓜多尔	巴拉圭	苏里南 洛杉矶 委内瑞拉	秘鲁	乌拉圭	巴西 华盛顿 圭亚那	
玻利维亚	智利	阿根廷

北美
安提瓜和巴布达	多米尼克	加拿大	圣卢西亚 巴巴多斯	格林纳达	美国	 美洲
圣文森特和格林纳丁斯 巴哈马	哥斯达黎加	墨西哥	特立尼达和多巴哥 巴拿马	
古巴	尼加拉瓜	危地马拉 伯利兹	海地	萨尔瓦多	牙买加 多米尼加	洪都拉斯	圣基茨和尼维斯

大洋洲
澳大利亚 澳洲 密克罗尼西亚	瓦努阿图	基里巴斯 新西兰	马绍尔群岛	图瓦卢	汤加 
巴布亚新几内亚	瑙鲁	斐济	库克群岛 帕劳	所罗门群岛	萨摩亚	纽埃

非洲
阿尔及利亚	刚果(布)	利比里亚	尼日利亚 埃及	刚果(金)	利比亚	塞拉利昂
埃塞俄比亚	吉布提	卢旺达	塞内加尔 安哥拉	几内亚	马达加斯加	塞舌尔
贝宁	几内亚比绍	马拉维	圣多美和普林西比 博茨瓦纳	加纳	马里	斯威士兰 布基纳法索	
加蓬	毛里求斯	苏丹 布隆迪	津巴布韦	毛里塔尼亚	索马里 赤道几内亚	喀麦隆	摩洛哥
坦桑尼亚 多哥	科摩罗	莫桑比克	突尼斯 厄立特里亚	科特迪瓦	纳米比亚	乌干达
佛得角	肯尼亚	南非	赞比亚 冈比亚	莱索托	南苏丹	乍得 尼日尔	中非

欧洲
阿尔巴尼亚	波兰	拉脱维亚	瑞士 爱尔兰	丹麦	立陶宛	塞尔维亚 爱沙尼亚 德国
列支敦士登	塞浦路斯 安道尔	俄罗斯	卢森堡	圣马力诺 奥地利	法国	罗马尼亚	
斯洛伐克 白俄罗斯	梵蒂冈	马耳他	斯洛文尼亚 保加利亚	芬兰	摩尔多瓦	乌克兰
北马其顿	荷兰	摩纳哥	西班牙 比利时	黑山	挪威	希腊 冰岛	捷克	葡萄牙	匈牙利
波黑	克罗地亚	瑞典	意大利 英国

亚洲 亚太
中国	哈萨克斯坦	缅甸	叙利亚 阿富汗	韩国	尼泊尔	亚美尼亚 阿联酋	吉尔吉斯斯坦	
日本	也门 阿曼	柬埔寨	沙特阿拉伯	伊拉克 阿塞拜疆	卡塔尔	斯里兰卡 兰卡 伊朗
巴基斯坦	科威特	塔吉克斯坦	以色列 巴勒斯坦	老挝	泰国 	印度 巴林	黎巴嫩	
土耳其	印度尼西亚 印尼 不丹	马尔代夫	土库曼斯坦	约旦 朝鲜	马来西亚 	文莱	
越南 东帝汶	蒙古	乌兹别克斯坦	格鲁吉亚 菲律宾	孟加拉国 孟加拉	新加坡

离岸 境外
香港 台湾 澳门 维尔京群岛 ‌巴拿马 ‌开曼群岛 开曼 百慕大 ‌卢森堡 塞浦路斯 加勒比海 

乌兰巴托 平壤 首尔 东京 远东 河内 万象 金边 仰光 曼谷 吉隆坡 新加坡 斯里巴加湾市 
马尼拉 雅加达 帝力 同奈

阿斯塔纳 比什凯克 塔什干 杜尚别 阿什哈巴德 喀布尔 伊斯兰堡 延布 新德里 达卡 
科伦坡 马累

德黑兰 巴格达 科威特城 利雅得 麦纳麦 多哈 阿布扎比 马斯喀特 萨那 安曼 拉马拉 
特拉维夫 大马士革 贝鲁特 安卡拉 尼科西亚 巴库 第比利斯 埃里温 迪拜

雷克雅未克 朗伊尔城 托尔斯港 哥本哈根 奥斯陆 斯德哥尔摩 赫尔辛基 玛丽港 塔林 里加
维尔纽斯

都柏林 伦敦 加的夫 爱丁堡 贝尔法斯特 圣彼得港 圣赫利尔 道格拉斯 巴黎 阿姆斯特丹
布鲁塞尔 卢森堡 柏林 维也纳 伯尔尼 瓦杜兹

里斯本 马德里 直布罗陀 安道尔 摩纳哥 罗马 梵蒂冈城 瓦莱塔 圣马力诺 雅典

莫斯科 明斯克 基辅 华沙 布拉格 布拉迪斯拉发 布达佩斯 卢布尔雅那 萨格勒布 萨拉热窝 
巴尼亚卢卡 贝尔格莱德 诺维萨德 普里什蒂纳 波德戈里察 斯科普里 地拉那 基希讷乌 
布加勒斯特 索非亚

开罗 喀土穆 的黎波里 突尼斯 阿尔及尔  拉巴特 西撒哈拉

努瓦克肖特 达喀尔 班珠尔 巴马科 瓦加杜古 普拉亚 比绍 科纳克里 弗里敦 蒙罗维亚
亚穆苏克罗 阿克拉 洛美 波多诺伏 尼亚美 阿布贾 詹姆士敦

雅温得 马拉博 恩贾梅纳 班吉 利伯维尔 布拉柴维尔 金沙萨 圣多美 罗安达

阿斯马拉 亚的斯亚贝巴 吉布提 摩加迪沙 内罗毕 坎帕拉 达累斯萨拉姆 基加利 布琼布拉 
利隆圭 马普托 塔那那利佛 维多利亚 莫罗尼 马穆楚 路易斯港 圣丹尼斯

卢萨卡 哈拉雷 哈博罗内 温得和克 比勒陀利亚 马塞卢 姆巴巴纳 维拉港 努美阿 瑙鲁
堪培拉 惠灵顿 莫尔兹比港 霍尼亚拉 苏瓦 塞班 阿加尼亚 科洛尼亚 科罗尔 马朱罗 拜里基 

富纳富提 马塔乌图 托克劳 阿皮亚 帕果帕果 努库阿洛法 阿洛菲 阿瓦鲁阿 帕皮提 
亚当斯敦 努克

""".split()

WW2 = ["澳門", "上纬(马)有限公司", "港湾(泰)有限公司", "力帆国际(控股)有限公司",
       "淡水河谷(国际)有限公司", "普力马(控股)有限公司", "兖煤国际(控股)有限公司",
       "汤佛(控股)有限公司", "亚森纳(控股)有限公司", "大士发电(私人)有限公司",
       "金螳螂(国际)建筑装饰有限公司"]

WW3 = """
耀晶有限公司 邦才有限公司 万科嘉宜控股有限公司 利亨有限公司 汇恩有限公司 
科修达有限公司 丽钻有限公司 发亮有限公司 瑞嘉投资实业有限公司 汇泰投资有限公司
招商局货柜服务有限公司 深美控股公司 立运有限公司 隆亨资本有限公司
金科物业服务集团有限公司 佩利雅有限公司 佩利雅公司 捷扬讯科国际有限公司
启航国际进出口有限公司 美的国际控股有限公司 世纪开利家用空调有限公司 
啟添投资有限公司 国际航空租赁有限公司 佰润国际有限公司 炼石投资有限公司
广发全球资本有限公司 泛广发展有限公司 美达尼龙有限公司 鸿愉有限公司 
九洲多媒体有限公司 才智商店有限公司 达峰国际股份有限公司 招商局置地有限公司
招商局置业有限公司 招商证券国际有限公司 招商局船务企业有限公司
招商轮船散货船控股有限公司 常隆有限公司 宝鹰国际投资控股有限公司
宝鹰国际建设投资有限公司 赛轮国际轮胎有限公司 软控(美洲)有限公司
软控美洲有限公司 金螳螂(国际)建筑装饰有限公司 金螳螂国际发展有限公司 
金智国际有限公司 欧畅国际控股有限公司 凯荟集团有限公司 洛美集装箱码头公司 
益光国际 格伦风电场有限公司 金风国际可再生能源有限公司 飞鸿国际发展有限公司
赛凌有限公司 潮宏基国际有限公司 顺便智能柜有限公司 迅传投资有限公司
妙机有限公司 順豐國際服務有限公司 順翔飛揚控股有限公司 順豐價值鏈科技服務有限公司
衡風有限公司 妙机有限公司 明迅投资有限公司 迅扬投资有限公司 迅翔投资有限公司
迅音投资有限公司 迅连投资有限公司 迅程投资有限公司 隽盈投资有限公司
達譽國際有限公司 陽帆控股有限公司 翠玉控股有限公司 紀澈有限公司 顺翔有限公司
呈润有限公司 嘉里企業方案有限公司 豐修科技有限公司 智得国际企业有限公司
沪士国际有限公司 常宝国际控股有限公司 锐岭集团有限公司 雅化国际投资发展有限公司
思望通讯有限公司 英飞拓国际有限公司 强劲国际工程有限公司 普尔德控股有限公司
捷航企业有限公司 奥瑞泰体育有限公司 健力有限公司 灿坤先端智能股份有限公司
蓝色光标国际传播集团有限公司 数码视讯国际有限公司 福瑞控股有限公司
澳洲佰盛有限公司 恒泰艾普石油集团有限公司 扬诚水务有限公司 苏交科国际有限公司
苏交科集团兰卡(私人)有限公司 汉鼎国际发展有限公司 明亚国际有限公司 
博腾美研有限公司 益信企业有限公司 联合好运集团控股有限公司 润欣勤增科技有限公司
润欣勤增科技股份有限公司 宸毅科技限公司 宸毅科技有限公司 鹏福发展有限公司
宏濑光电有限公司 泰博设计有限公司 奥飞数据国际有限公司 大士发电(私人)有限公司
寰宇船务企业有限公司 三一国际发展有限公司 银帆林业有限公司 大宋生物技术有限公司
同方国际有限公司 同方友友控股有限公司 泛广发展有限公司 恒裕通有限公司
金讯国际有限公司 兖煤澳洲附属公司 兖煤国际(控股)有限公司 普力马(控股)有限公司
汤佛(控股)有限公司 亚森纳(控股)有限公司 圆通国际控股有限公司 融诚林业股份有限公司
澳洲艾耐特照明有限公司 光涛投资有限公司 健民国际有限公司 施纳迪克国际公司
大橡塑国际有限公司 金地商置集团有限公司 士港科技有限公司 亨通光电国际有限公司
鹏欣国际集团有限公司 普霖强生生物制药有限公司 昂立国际投资有限公司
星科金朋私人有限公司 格隆希玛国际有限公司 裕璘实业有限公司 金迈拓国际有限公司
旭晨国际有限公司 明加尔金源公司 马钢瓦顿股份有限公司 圣灵投资银行 
杜克普百福工业股份公司 百福工业系统及机械有限公司 奇朔酒业有限公司 
白石风电场所有权有限公司 铁建宇翔有限公司 绿色动力投资控股有限公司 
怡球金属熔化有限公司 ﻿格飞驰航运有限公司 卡尔博气体净化技术有限公司
力帆国际(控股)有限公司 光大证券金融控股有限公司 港湾(泰)有限公司
凯程航运有限公司 凯志航运有限公司 凯锦航运有限公司 凯勇航运有限公司
淡水河谷(国际)有限公司 银河-联昌证券国际私人有限公司 江河幕墙美洲有限公司
诺顿金田有限公司 卡莫阿控股有限公司 卡莫阿铜业有限公司 金牌厨柜国际有限公司
金牌厨柜澳洲有限公司 太丞股份有限公司 骏亚国际电子有限公司 泰凌医药国际有限公司
莫克斯投资有限公司 火炬国际有限公司 贝能国际有限公司 上纬(马)有限公司

""".split()

World = WW1 + WW2 + WW3

mask2 = data_cn['GuaranteeName'].str.contains('|'.join(World))
filter1 = data_cn[mask2]      # with the World, maybe foreign

# Manual check companies
# filter2 = data_cn[~mask2]

# Concatenate the DataFrames
df2 = pd.concat([data_en, filter1], ignore_index=True)




df2.to_csv('240816_Event_level_data.csv', index=False, encoding='utf-8-sig')
